requirements.txt
langchain==0.0.205
gpt4all==0.3.4
chromadb==0.3.26
llama-cpp-python==0.1.64
urllib3==2.0.3
PyMuPDF==1.22.3
python-dotenv==1.0.0
unstructured==0.7.6
extract-msg==0.41.5
tabulate==0.9.0
pandoc==2.3
pypandoc==1.11
tqdm==4.65.0
auto_gptq==0.2.2
deep-translator==1.11.1
openai==0.27.8
huggingface==0.0.1
huggingface_hub==0.15.1
sentence_transformers==2.2.2
transformers==4.29.2
bitsandbytes==0.39.1
safetensors==0.3.1
torch==2.0.1
PyObjC==8.5; sys_platform == 'darwin' and platform_machine == 'x86_64'
fastapi==0.97.0
uvicorn==0.22.0
gunicorn==20.1.0
python-multipart==0.0.6
streamlit==1.23.1
streamlit-chat==0.0.2.2
streamlit-option-menu==0.3.6
psutil==5.9.5; sys_platform != 'darwin'
comtypes==1.2.0; sys_platform == 'linux' or sys_platform == 'linux2'  or sys_platform == 'linux3'
mammoth==1.6.0
ebooklib==0.18
beautifulsoup4==4.12.2
